Revert ":recycle: [ Refactor ] 페이지네이션 폴더구조 위계 넣기 및 기타 네이밍 변경"
This reverts commit 0b26737ad84ee3d108c1e7196dc6dae59cd2db9a.
