+++ src/components/organisms/RandomQuestionCard/useRandomQuestionCard.hook.jsx
@@ -2,17 +2,17 @@ import useRequestApi from '../../../hooks/useRequestApi';
 
 /**
  *
- * @returns randomCardData : object
+ * @returns randomCard : object
  */
 
 export default function useRandomQuestionCard() {
   const { data } = useRequestApi('subjects/?limit=100&offset=0', 'get');
 
-  let randomCardData;
+  let randomCard;
   if (data?.results?.length) {
     const randomIndex = Math.floor(Math.random() * data.results.length);
-    randomCardData = data.results[randomIndex];
+    randomCard = data.results[randomIndex];
   }
 
-  return { randomCardData };
+  return { randomCard };
 }
